S21: refactor(layout): 비율 저장을 f64 로 · spatial.rs 가 geometry.rs 경계로 이웃·순서를 정확히 판정한다
평평한 분할 렌더러 TRD §6 1b단계. 보통 레이아웃의 이웃·순서는 그대로이고, 아주 얇은 칸에서만 답이 옳아진다.

- LayoutNode::Split.ratio f32 → f64 — 받은 값이 그대로 돌아간다. LayoutNode.ts 는 바이트 동일(둘 다 number).
- spatial.rs 가 자기 사각형 계산(NormRect·assign_rects·leaf_rects)을 버리고 geometry::compute 의 경계를 입력으로 받는다.
- 인접 = 정확 비교(==) · 겹침 > 0 · 겹침 최대 후보 선택 유지. 절대 허용오차(EPS)·edge_eq·[EPS,1-EPS] 클램프를 걷었다 — 폭이 EPS 미만인 칸에서 이웃을 건너뛰거나 잃던 결함.
- 순서 = 중심점 f64::total_cmp + 동률은 트리 전위 순(안정 정렬).
- 면적 0 잎(테스트가 심은 값 등)은 이웃 판정·모서리 토큰에서 빠지고 순서 명단에는 남는다. 패닉 없음.
- 새 테스트 6건(얇은 칸 이웃 · 모서리 접촉 · 1 ulp 거짓 후보 · 면적 0 격리 · 모서리 토큰 · 결정성). geometry 의 spatial 교차 테스트는 대조 기준이 사라져 걷었다.

게이트: /review code full(codex PASS · doc-aware FIX 1회 — 앵커·주석 정정 뒤 PASS) · /qa standard PASS(워크스페이스 58 결과 줄 · 2762 통과 · lib_unit 284 · 프론트 971 · 바인딩 무변경).

+++ src-tauri/src/layout/types.rs
@@ -75,7 +75,8 @@ pub enum LayoutNode {
         #[ts(type = "string")]
         id: Uuid,
         dir: SplitDir,
-        ratio: f32,
+        // ADR-0227
+        ratio: f64,
         a: Box<LayoutNode>,
         b: Box<LayoutNode>,
     },
